ManagedOwnableContractWrapper introduced;
