+++ test/wrappers/fees_wrapper.js
@@ -23,83 +23,6 @@ describe('Fees Wrapper', () => {
   const exampleData = '0xda7a';
   const encodeAbiStub = sinon.stub().resolves(exampleData);
 
-  describe('getOwner', () => {
-    let getOwnerStub;
-    let getOwnerCallStub;
-    const ownerAddress = '0x1234ABCD';
-
-    before(async () => {
-      getOwnerStub = sinon.stub();
-      getOwnerCallStub = sinon.stub();
-      const contractMock = {
-        methods: {
-          owner: getOwnerStub.returns({
-            call: getOwnerCallStub.resolves(ownerAddress)
-          })
-        }
-      };
-      feesWrapper = new FeesWrapper();
-      sinon.stub(feesWrapper, 'contract').resolves(contractMock);
-    });
-
-    it('calls contract method with correct arguments', async () => {
-      await expect(feesWrapper.getOwner()).to.eventually.equal(ownerAddress);
-      expect(getOwnerStub).to.be.calledOnce;
-      expect(getOwnerCallStub).to.be.calledOnce;
-    });
-  });
-
-  describe('transferOwnership', () => {
-    const exampleAddress = '0xdeadbeef';
-    let transferOwnershipStub;
-    let transferOwnershipSendStub;
-    let contractMock;
-
-    before(async () => {
-      transferOwnershipStub = sinon.stub();
-      transferOwnershipSendStub = sinon.stub();
-      contractMock = {
-        methods: {
-          transferOwnership: transferOwnershipStub.returns({
-            send: transferOwnershipSendStub.resolves(),
-            encodeABI: encodeAbiStub
-          })
-        }
-      };
-    });
-
-    afterEach(() => {
-      resetHistory(contractMock);
-    });
-
-    describe('sendTransactions = true', () => {
-      before(() => {
-        feesWrapper = new FeesWrapper({}, {}, defaultAddress, true);
-        sinon.stub(feesWrapper, 'contract').resolves(contractMock);
-      });
-
-      it('calls contract method with correct arguments', async () => {
-        await feesWrapper.transferOwnership(exampleAddress);
-        expect(transferOwnershipStub).to.be.calledWith(exampleAddress);
-        expect(transferOwnershipSendStub).to.be.calledOnceWith({from: defaultAddress});
-      });
-    });
-
-    describe('sendTransactions = false', () => {
-      before(() => {
-        feesWrapper = new FeesWrapper({}, {}, defaultAddress, false);
-        sinon.stub(feesWrapper, 'contract').resolves(contractMock);
-      });
-
-      it('returns data', async () => {
-        expect(await feesWrapper.transferOwnership(exampleAddress)).to.equal(exampleData);
-        expect(transferOwnershipStub).to.be.calledWith(exampleAddress);
-        expect(transferOwnershipSendStub).to.be.not.called;
-        expect(encodeAbiStub).to.be.calledOnceWith();
-      });
-    });
-  });
-
   describe('setBaseUploadFee', () => {
     const exampleFee = '100';
     let contractMock;
